+++ docs/changelog.md
@@ -13,6 +13,7 @@ category: Administration
 
 * test: remove unused mock hints from test case and simplify test
 * docs: add developer documentation regarding unique simple name for actions (#115)
+* refactor: add unique prefix to active sessions action (#115)
 
 ### [20.06.0] - 2020-06-13